Create class for GithubCrawler and BatchPush
